Verify we can use an empty hash as an argument
Currently in 4.6.2 flog fails with a NoMethodError when attempting
to determine whether

    some_method({}) do |arg|
      # stuff
    end

is a dsl method or not. the dsl_name? method currently assumes that when
converted to sexp's the structure of the hash argument will always look
like:

    s(:hash, s(:lit, :woot), s(:lit, 42)))

but in the case above the structure is just `s(:hash)`.

This means that when we try and deep index into it using `[]`, we end up
calling `[]` on `nil` which raises.

This behaviour was fixed in master in seemingly unrelated commit
https://github.com/seattlerb/flog/commit/4f827fb59d4f44ca25d469e74a3e91d9547a9796

This commit introduces a regression test to confirm that the bahaviour
is now fixed.